python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openyourbubble"
version = "0.1.0"
description = "Local-first ingestion and study toolkit for OpenYourBubble"
authors = [{ name = "OpenYourBubble" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.9.0",
  "orjson>=3.9.10",
  "yake>=0.4.0",
  "feedparser>=6.0.11",
  "beautifulsoup4>=4.12.2",
  "lxml>=4.9.3",
  "readability-lxml>=0.8.1",
  "trafilatura>=1.6.1",
  "requests>=2.31.0",
  "networkx>=3.2.1"
]

[project.optional-dependencies]
llm = [
  "llama-cpp-python>=0.2.57"
]
translate = [
  "argostranslate>=1.9.6"
]

[project.scripts]
oyb = "openyourbubble.cli:app"
